refactor(settings): logging settings creating log folder
BUGFIXES

Previously, the log folder was inside src, which could result in a build with
an unwanted file if the application had been used before.

Now, the log folder is created when the application is configured, ensuring
that the executable file contains only folders and files necessary to run the
application, and the directories used by the application are created by the
application at run time.